+++ crates/codebook/src/queries/bash.scm
@@ -1,7 +1,7 @@
 (comment) @comment
 (string_content) @string
 (function_definition
-    name: (word) @identifier)
-(heredoc_body) @string
+    name: (word) @identifier.function)
+(heredoc_body) @string.heredoc
 (variable_assignment
-    name: (variable_name) @identifier)
+    name: (variable_name) @identifier.variable)