+++ crates/codebook/src/queries/css.scm
@@ -1,6 +1,6 @@
-(class_name) @identifier
-(id_name) @identifier
-(property_name) @identifier
+(class_name) @identifier.type
+(id_name) @identifier.type
+(property_name) @identifier.field
 (comment) @comment
 (string_value) @string
 (plain_value) @identifier